test(api): strengthen requesty abort assertions and share settle guard helper
